refactor(shared): move the SKILL.md serializer into @posthog/shared
serializeSkillMarkdown lived in workspace-server; move it to
@posthog/shared and have workspace-server re-export it. SKILL.md is a
serialization contract the cloud sandbox consumes, so any client that
emits skills (the workspace-server bundler today, a browser bundler
later) must produce byte-identical output — sharing one implementation
keeps them from drifting. No behavior change: desktop's bundler resolves
the same function through the re-export.

Generated-By: PostHog Code
Task-Id: 77d13a30-444d-4045-9d80-5e2a9f2e68ae
